1.4.5 리스트 메모 디테일 버전 구현
