+++ meme_android/app/src/main/java/com/likewhile/meme/data/model/MemoItem.kt
@@ -65,18 +65,3 @@ fun deserializeDrawingContent(data: ByteArray, filePath: String): String {
     return filePath
 }
 
-enum class MemoType(val typeValue: Int) {
-    TEXT(0),
-    LIST(1),
-    DRAWING(2);
-    companion object {
-        fun fromInt(value: Int): MemoType {
-            return when (value) {
-                1 -> LIST
-                2 -> DRAWING
-                else -> TEXT
-            }
-        }
-    }
-}
-
