+++ meme_android/app/src/main/java/com/likewhile/meme/data/model/MemoType.kt
@@ -1,0 +1,16 @@
+package com.likewhile.meme.data.model
+
+enum class MemoType(val typeValue: Int) {
+    TEXT(0),
+    LIST(1),
+    DRAWING(2);
+    companion object {
+        fun fromInt(value: Int): MemoType {
+            return when (value) {
+                1 -> LIST
+                2 -> DRAWING
+                else -> TEXT
+            }
+        }
+    }
+}
