+++ meme_android/app/src/main/java/com/likewhile/meme/ui/viewmodel/ListMemoViewModel.kt
@@ -5,6 +5,7 @@ import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MediatorLiveData
 import com.likewhile.meme.data.local.MemoDBHelper
+import com.likewhile.meme.data.model.ListItem
 import com.likewhile.meme.data.model.ListMemoItem
 import com.likewhile.meme.data.model.MemoItem
