+++ meme_android/app/src/main/res/layout/item_list.xml
@@ -3,7 +3,6 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
-
         <variable
             name="listItem"
             type="com.likewhile.meme.data.model.ListItem" />
