+++ meme_android/app/src/main/res/layout/item_memo_detail.xml
@@ -20,7 +20,7 @@
         <LinearLayout
             android:id="@+id/widget_memo"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
+            android:layout_height="match_parent"
             android:background="@color/white"
             android:orientation="vertical"
             android:padding="16dp">
@@ -57,8 +57,7 @@
                 android:layout_height="wrap_content"
                 android:text="@{memo.content}"
                 android:textColor="@color/black"
-                android:textSize="14sp"
-                android:maxHeight="200dp"/>
+                android:textSize="14sp"/>
 
             <TextView
                 android:id="@+id/textViewDate"
@@ -70,5 +69,4 @@
 
         </LinearLayout>
     </androidx.cardview.widget.CardView>
-
 </layout>
